Split tests to readmodel behaviors and projection behaviors
