+++ sample-app/src/test/scala/example/readmodel/rdb/DefaultConcertRepositorySpec.scala
@@ -1,6 +1,7 @@
 package example.readmodel.rdb
 
 import akka.actor.typed.ActorSystem
+import example.readmodel.rdb.projection.DatabaseConcertProjectionRepositoryBehaviors
 import example.readmodel.{ ConcertRepository, DefaultReadModelDiDesign }
 import testkit.AirframeDiSessionSupport
 import wvlet.airframe.Design
@@ -8,6 +9,7 @@ import wvlet.airframe.Design
 final class DefaultConcertRepositorySpec
     extends DatabaseConcertRepositorySpecBase()
     with DatabaseConcertRepositoryBehaviors
+    with DatabaseConcertProjectionRepositoryBehaviors
     with AirframeDiSessionSupport {
   override protected val design: Design =
     DefaultReadModelDiDesign.design
@@ -22,6 +24,7 @@ final class DefaultConcertRepositorySpec
 
   classOf[DefaultConcertRepository].getSimpleName should {
     behave like databaseConcertRepository(repository, databaseService)
+    behave like databaseConcertProjectionRepository(repository, databaseService)
   }
 
 }
